QR-codes containing *real* binary data now always work
In the first approach we assumed, that those QRs with that special
header (our default "#LSADxx"), which signals, that the rest of the QR
is binary data (which also can contain '\0's!), will always be coming
in multipart format.
Which isn't correct, all contents can also be in just one piece.
Now non-multipart-QRs also do work.